requirements.txt
ffmpeg-python>=0.2.0
streamlit>=1.28.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
yt-dlp>=2025.1.0